requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
faiss-cpu
langchain
langchain-community
pypdf
tiktoken
pocket-tts
torch
scipy

